+++ CHANGELOG.md
@@ -29,6 +29,8 @@
 * Invoice
   * Added the billing start and end dates to the invoice model (Thanks [@msize](https://github.com/msize))
   * Added the ReservationOrderId property to the one time invoice line item model
+  * Added the Quantity property to the daily rated usage line item model
+  * Addressed an issue that was preventing the pcToBCExchangeRate and pcToBCExchangeRateDate fields on the daily rated usage line item model from being deserialized
 * Subscription
   * Added the missing get subscription usage record by resource configuration (Thanks [@msize](https://github.com/msize))